Added Apache 2.0 hybrid license.
Creates a hybrid license for developers who wish to base the *Do No
Harm* license on the foundations of the *Apache 2.0* license.

The Apache license is clear in what you can/cannot do when it comes to
making derivatives of the license as found in the
[F.A.Q.](https://www.apache.org/foundation/license-faq.html#CLA-Usage)
Namely, the use of ASF and Apache must be removed.

The following details the changes and thought process used to craft the
Do No Harm/Apache 2.0 license:
 - The *preamble* was section was added first and then all sections
   were renumbered as necessary. This turned out to be a convenient
   introduction.
 - The *Forests* and *Deforestation* defintions also fit well within
   the Apache definitions section. BTW, adding a *Definitions* section
   to the core *Do No Harm* license might be worth consideration.
 - Only two of the *Redistribution* bullet points from the *Do No Harm*
   core were added to the *Redistribution* section as the other bullet
   points seemed to mimic what was in the core *Apache* license. BTW,
   adding a *Redistribution* section to the core *Do No Harm* license
   might be worth consideration.
 - The end of the license was updated to point back to the *Do No Harm*
   repository. This includes copyright modifications to fit the
   original *Apache* template.